LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/query-evaluation-expected.txt


FAIL (width) assert_equals: expected "true" but got "false"
PASS (height)
PASS (unknown)
PASS unknown(width)
FAIL ((width)) assert_equals: expected "true" but got "false"
PASS ((height))
PASS ((unknown))
FAIL ((((width)))) assert_equals: expected "true" but got "false"
PASS ((((height))))
PASS ((((unknown))))
PASS (not (width))
FAIL (not (height)) assert_equals: expected "true" but got "false"
PASS (not (unknown))
PASS (not unknown(width))
FAIL ((width) and (width)) assert_equals: expected "true" but got "false"
FAIL ((width) and (width) and (width)) assert_equals: expected "true" but got "false"
PASS ((height) and (height))
PASS ((height) and (width) and (width))
PASS ((width) and (height) and (width))
PASS ((width) and (width) and (height))
PASS ((unknown) and (width) and (width))
PASS ((width) and (unknown) and (width))
PASS ((width) and (width) and (unknown))
FAIL ((width) or (width)) assert_equals: expected "true" but got "false"
FAIL ((width) or (width) or (width)) assert_equals: expected "true" but got "false"
PASS ((height) or (height))
FAIL ((height) or (width) or (width)) assert_equals: expected "true" but got "false"
FAIL ((width) or (height) or (width)) assert_equals: expected "true" but got "false"
FAIL ((width) or (width) or (height)) assert_equals: expected "true" but got "false"
PASS ((unknown) or (width) or (width))
PASS ((width) or (unknown) or (width))
PASS ((width) or (width) or (unknown))
PASS ((unknown) or (height) or (width))
PASS (not ((width) and (width)))
FAIL (not ((width) and (height))) assert_equals: expected "true" but got "false"
PASS ((width) and (not ((height) or (width))))
FAIL ((height) or (not ((height) and (width)))) assert_equals: expected "true" but got "false"
PASS ((height) or ((height) and (width)))
